pyproject.toml
[tool.poetry]
name = "package-manager"
version = "0.1.0"
description = ""
authors = ["Ahmad Sadeghi <ahmad.81.sadeghi@gmail.com>"]
readme = "README.md"
packages = [{include = "package_manager"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-jalali-date = "^1.0.1"
openpyxl = "^3.1.2"
gunicorn = "^20.1.0"
whitenoise = "^6.5.0"
jdatetime = "^4.1.1"
djoser = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
